Create QR code generator
It imports the qrcode library.
The function generate_qr_code takes two arguments: data (the information you want to encode) and filename (the name you want to give to the generated image).
The function uses the qrcode.QRCode class to create a QR code with the specified data.
It then saves the QR code as an image file.
To run this script, you can save it as a .py file and execute it using a Python interpreter. When prompted, enter the data you want to encode and the filename for the generated QR code.

For example, if you enter "https://www.example.com" as the data and "example_qr_code.png" as the filename, it will generate a QR code that links to https://www.example.com and save it as "example_qr_code.png" in the same directory as your script.